refactor: fix lint
